combine limit and sort
